+++ My project1/Assets/animation/Free/Terrain/Rectangular Palette.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 4496cb665b1967e4a9b0a88aef47e639
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
