+++ My project1/Assets/prefab/player_6.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 7629cdedde833444c84632d83c051b6f
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
