+++ My project1/Assets/scrip/Character.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d6e6d6c0d5deadf41a5f74c967e19513
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
