+++ My project1/Assets/scrip/Character/Animation.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d1758fb26c90e054a96bbbe76b5935fa
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
