+++ My project1/Assets/scrip/Character/Animation/Move.anim
@@ -1,0 +1,83 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!74 &7400000
+AnimationClip:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_Name: Move
+  serializedVersion: 6
+  m_Legacy: 0
+  m_Compressed: 0
+  m_UseHighQualityCurve: 1
+  m_RotationCurves: []
+  m_CompressedRotationCurves: []
+  m_EulerCurves: []
+  m_PositionCurves: []
+  m_ScaleCurves: []
+  m_FloatCurves: []
+  m_PPtrCurves:
+  - curve:
+    - time: 0
+      value: {fileID: -1647304486, guid: e73b11be6be690b438f0ac532489b06e, type: 3}
+    - time: 0.21666667
+      value: {fileID: -679083876, guid: e73b11be6be690b438f0ac532489b06e, type: 3}
+    - time: 0.43333334
+      value: {fileID: 705851313, guid: e73b11be6be690b438f0ac532489b06e, type: 3}
+    - time: 0.65
+      value: {fileID: -997018097, guid: e73b11be6be690b438f0ac532489b06e, type: 3}
+    - time: 0.8666667
+      value: {fileID: 1128159921, guid: e73b11be6be690b438f0ac532489b06e, type: 3}
+    - time: 1.0833334
+      value: {fileID: 117443887, guid: e73b11be6be690b438f0ac532489b06e, type: 3}
+    attribute: m_Sprite
+    path: 
+    classID: 212
+    script: {fileID: 0}
+  m_SampleRate: 60
+  m_WrapMode: 0
+  m_Bounds:
+    m_Center: {x: 0, y: 0, z: 0}
+    m_Extent: {x: 0, y: 0, z: 0}
+  m_ClipBindingConstant:
+    genericBindings:
+    - serializedVersion: 2
+      path: 0
+      attribute: 0
+      script: {fileID: 0}
+      typeID: 212
+      customType: 23
+      isPPtrCurve: 1
+    pptrCurveMapping:
+    - {fileID: -1647304486, guid: e73b11be6be690b438f0ac532489b06e, type: 3}
+    - {fileID: -679083876, guid: e73b11be6be690b438f0ac532489b06e, type: 3}
+    - {fileID: 705851313, guid: e73b11be6be690b438f0ac532489b06e, type: 3}
+    - {fileID: -997018097, guid: e73b11be6be690b438f0ac532489b06e, type: 3}
+    - {fileID: 1128159921, guid: e73b11be6be690b438f0ac532489b06e, type: 3}
+    - {fileID: 117443887, guid: e73b11be6be690b438f0ac532489b06e, type: 3}
+  m_AnimationClipSettings:
+    serializedVersion: 2
+    m_AdditiveReferencePoseClip: {fileID: 0}
+    m_AdditiveReferencePoseTime: 0
+    m_StartTime: 0
+    m_StopTime: 1.1
+    m_OrientationOffsetY: 0
+    m_Level: 0
+    m_CycleOffset: 0
+    m_HasAdditiveReferencePose: 0
+    m_LoopTime: 1
+    m_LoopBlend: 0
+    m_LoopBlendOrientation: 0
+    m_LoopBlendPositionY: 0
+    m_LoopBlendPositionXZ: 0
+    m_KeepOriginalOrientation: 0
+    m_KeepOriginalPositionY: 1
+    m_KeepOriginalPositionXZ: 0
+    m_HeightFromFeet: 0
+    m_Mirror: 0
+  m_EditorCurves: []
+  m_EulerEditorCurves: []
+  m_HasGenericRootTransform: 0
+  m_HasMotionFloatCurves: 0
+  m_Events: []
