+++ My project1/Assets/scrip/Character/Animation/Move.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0d19e10c3d0a0a940913667b79c01fe4
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
