+++ My project1/Assets/scrip/Character/Animation/New Animation.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 84619098024b43e4196707e4637646dd
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
