+++ My project1/Assets/scrip/Character/Animation/Player.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 908a7188a500588488db7215fc5d83bd
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
