+++ My project1/Assets/scrip/Character/Animation/player_6.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c99ed178a3c5ebd4499871283f47f865
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
