+++ My project1/Assets/scrip/Character/playerMove.cs.meta
