+++ My project1/UserSettings/EditorUserSettings.asset
@@ -9,10 +9,10 @@ EditorUserSettings:
       value: 070803555c570a0a0b5d592015725d44404f4e2c7c2b20667e2a4f61b0b03568
       flags: 0
     RecentlyUsedSceneGuid-1:
-      value: 515250075c0c595e5f5a5e71122159444e4e4a2f7a7d7f602f284d66b4b76661
+      value: 0009575504015b0c585c582044255d444f154d2b747b276378704b67b2b1646f
       flags: 0
     RecentlyUsedSceneGuid-2:
-      value: 0009575504015b0c585c582044255d444f154d2b747b276378704b67b2b1646f
+      value: 515250075c0c595e5f5a5e71122159444e4e4a2f7a7d7f602f284d66b4b76661
       flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
